AI Resume Builder/requirements.txt
fastapi
openai
pydantic
python-dotenv
uvicorn